refactor: Update data handling to use Graph schema across multiple modules
